services/classifier-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
redis==5.0.1
openai==1.3.0
google-generativeai==0.3.0
aiohttp==3.9.0
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0